refactor(shared): move Course model to :shared module
